// CourseApp/client/src/setupTests.ts
// jest-dom adds custom jest matchers for asserting on DOM nodes.
// allows you to do things like:
// expect(element).toHaveTextContent(/react/i)
// learn more: https://github.com/testing-library/jest-dom
import '@testing-library/jest-dom';
export interface Course {
    id: number;
    name: string;
    lessons: number;
    startDate: string;
    completedLessons: number;
    duration: string;
  }
  